I edited:
Player -  added a constructor with one parameter, String username.

CreatePlayerActivity - Added code on the OnCreate method to interact with
the TextField.
OnAddPressed still needs to be implemented so we can actually use the
data passed in by the user.

activity_player_creation - Edited the layout so now the user can enter a
username.